+++ application/wypl-core/src/main/java/com/wypl/wyplcore/schedule/data/response/ScheduleInfoCreateResponse.java
@@ -1,0 +1,14 @@
+package com.wypl.wyplcore.schedule.data.response;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public record ScheduleInfoCreateResponse (
+
+        @JsonProperty("calendar_id")
+        Long calendarId,
+
+        @JsonProperty("schedule_info_id")
+        Long scheduleInfoId
+
+){
+}
