+++ domain/jpa-calendar-domain/src/main/java/com/wypl/jpacalendardomain/calendar/mapper/ScheduleInfoMapper.java
@@ -1,17 +1,14 @@
 package com.wypl.jpacalendardomain.calendar.mapper;
 
-
-import com.wypl.jpacalendardomain.calendar.data.ConvertibleScheduleInfo;
 import com.wypl.jpacalendardomain.calendar.domain.Calendar;
 import com.wypl.jpacalendardomain.calendar.domain.ScheduleInfo;
 
 public class ScheduleInfoMapper {
 
-    public static ScheduleInfo toJpaScheduleInfo(ConvertibleScheduleInfo convertibleScheduleInfo, Calendar calendar) {
+    public static ScheduleInfo toJpaScheduleInfo(Calendar calendar, long creatorId) {
         return ScheduleInfo.builder()
-                .creatorId(convertibleScheduleInfo.getCreatorId())
+                .creatorId(creatorId)
                 .calendar(calendar)
                 .build();
     }
-
 }
